Add support for references to macro parameters
Implements a jerry-rigged way of resolving references to macro
parameters by substituting the element type of the parameter for the
same element type represented by the type keyword.  I.E., when a
parameter with a keyword of 'type' is found, the processor will treat it
as a top level type declaration when processing declarations in that
scope.